final Update

+++ src/Pages/User/MyOrders.js
@@ -6,29 +6,31 @@ const MyOrders = () => {
     const [orders, setOrders] = useState([])
     const { user } = UseAuth()
     useEffect(() => {
-        fetch(`https://young-basin-54611.herokuapp.com/myOrder/${user?.email}`)
+        fetch(`http://localhost:5000/myOrder/${user?.email}`)
             .then((res) => res.json())
             .then((data) => setOrders(data));
     }, [user?.email]);
 
   
     const handleDeleteUserOrder = id => {
-      const url = `https://young-basin-54611.herokuapp.com/allOrders/${id}`;
-        fetch(url, {
-            method: 'DELETE'
-  
-        })
-            .then(res => res.json())
-            .then(data => {
-             
-                if (data.deletedCount) {
-                    const remaining = orders?.filter(order => order._id !== id);
-                    setOrders(remaining);
-                    alert('Remove Your Order Successfully')
-  
-                }
-  
-            })
+      const process = window.confirm('Are you sure you want to Cancel this Order')
+
+     if(process){
+      const url = `http://localhost:5000/allOrders/${id}`;
+      fetch(url, {
+          method: 'DELETE'
+      })
+          .then(res => res.json())
+          .then(data => {
+           
+              if (data.deletedCount) {
+                  const remaining = orders?.filter(order => order._id !== id);
+                  setOrders(remaining);
+
+              }
+
+          })
+     }
   
     }
 
@@ -38,8 +40,8 @@ const MyOrders = () => {
 <h2 className="  AddServiceHeader p-3  mx-auto mt-1"> You Have {orders.length} Order</h2>
 
 <table role="table" style={{width: "100%"}}>
-<thead role="rowgroup" >
-    <tr role="row " className="bg-dark text-white mb-3 p-2" style={{border: "1px solid red"}}>
+<thead  >
+    <tr className="bg-dark text-white mb-3 p-2" style={{border: "1px solid red"}}>
       <th role="columnheader" >Product</th>
       <th role="columnheader">Price</th>
       <th role="columnheader">Status</th>
@@ -47,7 +49,7 @@ const MyOrders = () => {
     </tr>
   </thead>
   {orders?.map((order) => (
-     <tbody role="rowgroup">
+     <tbody  key={order._id}>
      <tr role="row" style={{border: "2px solid gray"}} >
        <td role="cell">{order.name}</td>
        <td role="cell">{order.price}</td>
